Naive implementation of padding for the :packed layout.
We simply add padding to the bottom and right of every image,
without trying to figure out if we are at the edge of the image
or not. This solves my own need for padding: getting rid of
display bugs on touch devices due to scaling (zooming) the
document. This is resolved by adding 1px of padding around all
images.